+++ api/src/models/units/Unit.ts
@@ -13,7 +13,7 @@ import {IProgress} from '../../../../shared/models/progress/IProgress';
 import {User} from '../User';
 
 interface IUnitModel extends IUnit, mongoose.Document {
-  exportJSON: () => Promise<IUnit>;
+  exportJSON: (onlyBasicData?: boolean) => Promise<IUnit>;
   calculateProgress: (users: IUser[], progress: IProgress[]) => Promise<IUnit>;
   populateUnit: () => Promise<IUnitModel>;
   secureData: (user: IUser) => Promise<IUnitModel>;
@@ -58,8 +58,13 @@ const unitSchema = new mongoose.Schema({
     toObject: {
       virtuals: true,
       transform: function (doc: IUnitModel, ret: any) {
-        ret._id = ret._id.toString();
-        ret._course = ret._course.toString();
+        if (ret._id) {
+          ret._id = ret._id.toString();
+        }
+
+        if (ret._course) {
+          ret._course = ret._course.toString();
+        }
       }
     },
   }
@@ -72,7 +77,7 @@ unitSchema.virtual('progressData', {
   justOne: true
 });
 
-unitSchema.methods.exportJSON = function () {
+unitSchema.methods.exportJSON = function (onlyBasicData: boolean= false) {
   const obj = this.toObject();
 
   // remove unwanted informations
@@ -84,8 +89,15 @@ unitSchema.methods.exportJSON = function () {
   delete obj.unitCreator;
   delete obj.files;
 
+
   // custom properties
   delete obj._course;
+
+  if (onlyBasicData) {
+    delete obj.id;
+    delete obj.progressData;
+  }
+
   return obj;
 };